feat: set soft and hard limits for CPU time
This PR modifies how the CPU time limits are enforced. Instead of
accounting for bursts, we do set soft and a hard limit. When the soft limit
is reached the worker is retired (ie. current request completes, but no
new requests are accepted). And worker is terminated only when it hits
the hard limit. This should reduce the chances of function failing
mid-request due to CPU limits.