refactor: internal unix streams must have a synchronized connection lifetime with an outbound request
Currently, The unix stream pair is used to communicate with the main thread,
main worker, and user workers.

But each stream pair is independent of the other pair, so there are no
guarantees that it will be synchronized between stream pairs.

If streams are not synchronized, each stream will have a different lifetime, so
depending on the case, it will cause unexpected connection breakage.

This problem was already exposed as an internal server error or timeout on the
surface, which was too hard to notice.

It might be a substantial solution for the Pull Requests and Issues described
below.

supabase/edge-runtime#50supabase/edge-runtime#56supabase/edge-runtime#69supabase/edge-runtime#75
supabase/edge-runtime#198

🫠

(cherry picked from commit 2387c9eb720d07b7de8bff4e0852048e24bd12b2)