feat: introduce `oneshot` supervisor policy
This policy forces the supervisor to terminate the isolation immediately if the
request is complete.

Using this policy with development will make sense because it terminates the
isolation immediately if the request is complete, so developers will not have
to restart runtime.

This commit solves cases such as supabase/edge-runtime#192 and supabase/edge-runtime#212

(cherry picked from commit 0b1ddd0865991b2449dfebfaac6c834dbbc0a7ce)