fix: don't leave pending requests orphan state for a long time
If the spawned user workers go to a retired state at once, then there will be no
worker to accept the pending requests for a long time. So we must bring out
their semaphore permits while going to a retired state so the pooler can create
a fresh user worker.