refactor(cpu_timer): rework the handle signal part
Handling the Unix signals to be stable in the multi-thread environment without
writing a lot of code is nearly impossible because there are a lot of pitfalls
in there. Luckily, we already know the crates that can solve them. This commit
refactors the signal handling part into the tokio message queue-based handling
using the `signal-hook` crates.