fix: what a silly mistake again 🫠
We should call `Isolate::Enter` because we must bind the isolate to the thread
when evaluating the mod initially, but that doesn't mean it should only be
bound when evaluating. Naturally, we should invoke `Isolate::Enter` even when
the v8 isolate enters its loop again.

What a silly mistake again 😫