fix: thread safe handle may no longer have an isolate pointer
There is no explicit lifetime synchronization path between the supervisor and
isolate, so we cannot assume that `Isolate::request_interrupt` will always be
scheduled.

The `IsolateInterruptData` passed as an argument has always been passed to
`Box::into_raw`, so if the isolation is ever destroyed, the step of converting
it back to a `Box<T>` to free the memory is omitted. We should have looked at
the result of the `request_interrupt` call and freed the passed arguments in
some cases, but we didn't, which led to a memory leak.

(cherry picked from commit 2790fa44e609470584e521fdb3fa5137f8583692)