fix(base): give the runtime an explicit chance to handle an interruption request by the supervisor
DESCRIPTION:

If we cancel the `termination_request_token` token too early, `DenoRuntime`
loses the chance to handle the interruption and may not provide a value for
`isolate_memory_usage_rx`, causing the task that called `create_supervisor` to
deadlock.

`DenoRuntime::run` is returned by canceling the `termination_request_token`
token, so the user will see the worker terminated as normal, but it will be
reported as `EventLoopCompleted`.